--- spyder/app/tests/spyder_boilerplate/spyder/api.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# ----------------------------------------------------------------------------
-# Copyright © 2021, Spyder Bot
-#
-# Licensed under the terms of the MIT license
-# ----------------------------------------------------------------------------
-"""
-Spyder Boilerplate API.
-"""
