+++ spyder/app/tests/spyder_boilerplate/spyder/plugin.py
@@ -8,19 +8,101 @@
 Spyder Boilerplate Plugin.
 """
 
-# Third-party imports
+# Third party imports
 from qtpy.QtGui import QIcon
+from qtpy.QtWidgets import QHBoxLayout, QLabel
 
 # Spyder imports
+from spyder.api.config.decorators import on_conf_change
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
+from spyder.api.preferences import PluginConfigPage
 from spyder.api.translations import get_translation
+from spyder.api.widgets.main_widget import PluginMainWidget
 from spyder.plugins.layout.layouts import VerticalSplitLayout2
 
-# Local imports
-from spyder_boilerplate.spyder.confpage import SpyderBoilerplateConfigPage
-from spyder_boilerplate.spyder.widgets import SpyderBoilerplateWidget
+_ = lambda text: text
 
-_ = get_translation("spyder_boilerplate.spyder")
+
+class SpyderBoilerplateConfigPage(PluginConfigPage):
+
+    # --- PluginConfigPage API
+    # ------------------------------------------------------------------------
+    def setup_page(self):
+        pass
+
+
+class SpyderBoilerplateActions:
+    ExampleAction = "example_action"
+
+
+class SpyderBoilerplateToolBarSections:
+    ExampleSection = "example_section"
+
+
+class SpyderBoilerplateOptionsMenuSections:
+    ExampleSection = "example_section"
+
+
+class SpyderBoilerplateWidget(PluginMainWidget):
+
+    # PluginMainWidget class constants
+
+    # Signals
+
+    def __init__(self, name=None, plugin=None, parent=None):
+        super().__init__(name, plugin, parent)
+
+        # Create an example label
+        self._example_label = QLabel("Example Label")
+
+        # Add example label to layout
+        layout = QHBoxLayout()
+        layout.addWidget(self._example_label)
+        self.setLayout(layout)
+
+    # --- PluginMainWidget API
+    # ------------------------------------------------------------------------
+    def get_title(self):
+        return _("Spyder Boilerplate")
+
+    def get_focus_widget(self):
+        pass
+
+    def setup(self):
+        # Create an example action
+        example_action = self.create_action(
+            name=SpyderBoilerplateActions.ExampleAction,
+            text="Example action",
+            tip="Example hover hint",
+            icon=self.create_icon("spyder"),
+            triggered=lambda: print("Example action triggered!"),
+        )
+
+        # Add an example action to the plugin options menu
+        menu = self.get_options_menu()
+        self.add_item_to_menu(
+            example_action,
+            menu,
+            SpyderBoilerplateOptionsMenuSections.ExampleSection,
+        )
+
+        # Add an example action to the plugin toolbar
+        toolbar = self.get_main_toolbar()
+        self.add_item_to_toolbar(
+            example_action,
+            toolbar,
+            SpyderBoilerplateOptionsMenuSections.ExampleSection,
+        )
+
+    def update_actions(self):
+        pass
+
+    @on_conf_change
+    def on_section_conf_change(self, section):
+        pass
+
+    # --- Public API
+    # ------------------------------------------------------------------------
 
 
 class SpyderBoilerplate(SpyderDockablePlugin):
@@ -33,6 +115,7 @@ class SpyderBoilerplate(SpyderDockablePlugin):
     OPTIONAL = []
     WIDGET_CLASS = SpyderBoilerplateWidget
     CONF_SECTION = NAME
+    CONF_WIDGET_CLASS = SpyderBoilerplateConfigPage
     CUSTOM_LAYOUTS = [VerticalSplitLayout2]
 
     # --- Signals
